Restore float gauge boundaries in PromQL
Keep gauges narrowed at GaugeScan ingestion boundaries while retaining double scalar calculations where appropriate. Use stable online variance and compensated aggregate accumulation, and add a focused single-suite integration runner for regression isolation.

+++ okapi-core/src/main/java/org/okapi/metrics/query/QueryRecords.java
@@ -12,5 +12,5 @@ public class QueryRecords {
   public record Slice(
       String series, long from, long to, RES_TYPE resolution, AGG_TYPE aggregation) {}
 
-  public record QueryResult(String name, List<Long> timestamps, List<Double> values) {}
+  public record QueryResult(String name, List<Long> timestamps, List<Float> values) {}
 }
